requirements.txt
mido==1.3.2
paho-mqtt==1.6.1
